tests/wpt/web-platform-tests/html/canvas/element/drawing-text-to-the-canvas/2d.text.drawing.style.font-relative.spacing.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>Canvas test: 2d.text.drawing.style.font-relative.spacing</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.text.drawing.style.font-relative.spacing</h1>
<p class="desc">Testing letter spacing and word spacing with font-relative length</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("Testing letter spacing and word spacing with font-relative length");
_addTest(function(canvas, ctx) {

  _assertSame(ctx.letterSpacing, '0px', "ctx.letterSpacing", "'0px'");
  _assertSame(ctx.wordSpacing, '0px', "ctx.wordSpacing", "'0px'");

  ctx.letterSpacing = '1EX';
  ctx.wordSpacing = '1EM';
  _assertSame(ctx.letterSpacing, '1ex', "ctx.letterSpacing", "'1ex'");
  _assertSame(ctx.wordSpacing, '1em', "ctx.wordSpacing", "'1em'");

  ctx.letterSpacing = '1ch';
  ctx.wordSpacing = '1ic';
  _assertSame(ctx.letterSpacing, '1ch', "ctx.letterSpacing", "'1ch'");
  _assertSame(ctx.wordSpacing, '1ic', "ctx.wordSpacing", "'1ic'");

});
</script>
